refactor - moved posts form to top of home page, removed admin posts page
