+++ src/pages/admin/index.tsx
@@ -3,8 +3,6 @@ import { Typography } from "@mui/material";
 import { NextPage } from "next";
 import { isLoggedInVar } from "../../client/cache";
 import LevelOneHeading from "../../components/Shared/LevelOneHeading";
-import Link from "../../components/Shared/Link";
-import { NavigationPaths } from "../../constants/common";
 import { useTranslate } from "../../hooks/common";
 
 // TODO: Add remaining admin functionality and views - below is a WIP
@@ -18,22 +16,11 @@ const Admin: NextPage = () => {
 
   return (
     <>
-      <LevelOneHeading style={{ fontSize: 18, marginBottom: 50 }}>
+      <LevelOneHeading style={{ fontSize: 18, marginBottom: 40 }}>
         {t("admin.labels.admin")}
       </LevelOneHeading>
 
       <Typography gutterBottom>{t("prompts.wip")}</Typography>
-
-      <Typography>
-        Posts can be managed{" "}
-        <Link
-          href={NavigationPaths.AdminPosts}
-          style={{ fontWeight: "bold", color: "black" }}
-        >
-          here
-        </Link>
-        .
-      </Typography>
     </>
   );
 };
